requirements.txt
asgiref==3.8.1
Django==4.2.5
django-cors-headers==4.7.0
djangorestframework==3.14.0
pytz==2025.1
setuptools==75.8.0
sqlparse==0.5.3
wheel==0.45.1
